dead code deleted, config moved to types
