+++ packages/js/src/account/account-manager.ts
@@ -4,11 +4,10 @@ import * as process from 'process';
 import * as nearAPI from 'near-api-js';
 import {NEAR} from 'near-units';
 import {asId, isTopLevelAccount, randomAccountId, timeSuffix} from '../utils';
-import {KeyPair, BN, KeyPairEd25519, FinalExecutionOutcome, KeyStore, AccountBalance, NamedAccount, PublicKey, AccountView} from '../types';
+import {Config, KeyPair, BN, KeyPairEd25519, FinalExecutionOutcome, KeyStore, AccountBalance, NamedAccount, PublicKey, AccountView} from '../types';
 import {debug, txDebug} from '../internal-utils';
 import {Transaction} from '../transaction';
 import {JsonRpcProvider} from '../jsonrpc';
-import {Config} from '../interfaces';
 import {TransactionResult} from '../transaction-result';
 import {Account} from './account';
 import {NearAccount} from './near-account';
