+++ packages/js/src/account/near-account-manager.ts
@@ -1,10 +1,9 @@
 import {KeyPair} from 'near-api-js';
 import {NEAR} from 'near-units';
 import {TransactionResult} from '../transaction-result';
-import {Config} from '../interfaces';
 import {JsonRpcProvider} from '../jsonrpc';
 import {Transaction} from '../transaction';
-import {AccountBalance, AccountView} from '../types';
+import {Config, AccountBalance, AccountView} from '../types';
 import {NearAccount} from './near-account';
 
 export interface NearAccountManager {
