+++ packages/js/src/index.ts
@@ -11,5 +11,4 @@ export * from './types';
 export * from './account';
 export * from './transaction-result';
 export * from './jsonrpc';
-export * from './interfaces';
 export * from 'near-units';
